requirements.txt
crawl4ai==0.8.6
pydantic==2.13.3
playwright==1.58.0
python-dotenv==1.2.2
